// Frontend/src/enums/CM/cm007.ts
export enum Cm007Status {
  All = 'All',
  Draft = 'Draft',
  Editing = 'Editing',
  Rejected = 'Rejected',
  WaitingCommitteeApproval = 'WaitingCommitteeApproval',
  WaitingAssignment = 'WaitingAssignment',
  WaitingComment = 'WaitingComment',
  WaitingApproval = 'WaitingApproval',
  Approved = 'Approved',
  RejectedToAssignee = 'RejectedToAssignee',
  WaitingAddendumAssignment = 'WaitingAddendumAssignment',
  WaitingDraftAddendum = 'WaitingDraftAddendum',
  WaitingReview = 'WaitingReview',
};

export enum Cm007AccordionTab {
  Committee = 'Committee',
  Assignee = 'Assignee',
  Acceptor = 'Acceptor',
  AddendumDrafter = 'AddendumDrafter',
  Reviewer = 'Reviewer',
};